+++ inst/include/dtwclust.hpp
@@ -15,6 +15,14 @@ SEXP dba(SEXP X, SEXP centroid,
     return fun(X, centroid, max_iter, delta, trace, multivariate, mv_ver, DOTS);
 }
 
+SEXP dtw_lb(SEXP X, SEXP Y, SEXP D, SEXP DOTS)
+{
+    static SEXP(*fun)(SEXP, SEXP, SEXP, SEXP) = NULL;
+    if(fun == NULL)
+        fun = (SEXP(*)(SEXP, SEXP, SEXP, SEXP)) R_GetCCallable("dtwclust", "dtw_lb");
+    return fun(X, Y, D, DOTS);
+}
+
 SEXP envelope(SEXP series, SEXP window)
 {
     static SEXP(*fun)(SEXP, SEXP) = NULL;