+++ src/dtwclustpp.h
@@ -9,6 +9,8 @@ RcppExport SEXP dba(SEXP X, SEXP centroid,
                     SEXP max_iter, SEXP delta, SEXP trace,
                     SEXP multivariate, SEXP mv_ver, SEXP DOTS);
 
+RcppExport SEXP dtw_lb(SEXP X, SEXP Y, SEXP D, SEXP DOTS);
+
 RcppExport SEXP envelope(SEXP series, SEXP window);
 
 void envelope_cpp(const Rcpp::NumericVector& array, unsigned int width,