+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ## Version 4.0.1
+* Ported `dtw_lb` to `C++` when using `dtw_basic`.
 
 ## Version 4.0.0
 * Optimized `TADPole` for multiple `k` and `dc` values.